move test to mod
