+++ zebrad/src/components/sync.rs
@@ -681,57 +681,3 @@ where
         );
     }
 }
-
-#[cfg(test)]
-mod test {
-    use zebra_chain::parameters::NetworkUpgrade;
-
-    use super::*;
-
-    /// Make sure the timeout values are consistent with each other.
-    #[test]
-    fn ensure_timeouts_consistent() {
-        zebra_test::init();
-
-        // This constraint clears the download pipeline during a restart
-        assert!(
-            SYNC_RESTART_DELAY.as_secs() > 2 * BLOCK_DOWNLOAD_TIMEOUT.as_secs(),
-            "Sync restart should allow for pending and buffered requests to complete"
-        );
-
-        // This constraint avoids spurious failures due to block retries timing out.
-        // We multiply by 2, because the Hedge can wait up to BLOCK_DOWNLOAD_TIMEOUT
-        // seconds before retrying.
-        const BLOCK_DOWNLOAD_HEDGE_TIMEOUT: u64 =
-            2 * BLOCK_DOWNLOAD_RETRY_LIMIT as u64 * BLOCK_DOWNLOAD_TIMEOUT.as_secs();
-        assert!(
-            SYNC_RESTART_DELAY.as_secs() > BLOCK_DOWNLOAD_HEDGE_TIMEOUT,
-            "Sync restart should allow for block downloads to time out on every retry"
-        );
-
-        // This constraint avoids spurious failures due to block download timeouts
-        assert!(
-            BLOCK_VERIFY_TIMEOUT.as_secs()
-                > SYNC_RESTART_DELAY.as_secs()
-                    + BLOCK_DOWNLOAD_HEDGE_TIMEOUT
-                    + BLOCK_DOWNLOAD_TIMEOUT.as_secs(),
-            "Block verify should allow for a block timeout, a sync restart, and some block fetches"
-        );
-
-        // The minimum recommended network speed for Zebra, in bytes per second.
-        const MIN_NETWORK_SPEED_BYTES_PER_SEC: u64 = 10 * 1024 * 1024 / 8;
-
-        // This constraint avoids spurious failures when restarting large checkpoints
-        assert!(
-            BLOCK_VERIFY_TIMEOUT.as_secs() > SYNC_RESTART_DELAY.as_secs() + 2 * zebra_consensus::MAX_CHECKPOINT_BYTE_COUNT / MIN_NETWORK_SPEED_BYTES_PER_SEC,
-            "Block verify should allow for a full checkpoint download, a sync restart, then a full checkpoint re-download"
-        );
-
-        // This constraint avoids spurious failures after checkpointing has finished
-        assert!(
-            BLOCK_VERIFY_TIMEOUT.as_secs()
-                > 2 * NetworkUpgrade::Blossom.target_spacing().num_seconds() as u64,
-            "Block verify should allow for at least one new block to be generated and distributed"
-        );
-    }
-}
